backend/llm_gateway/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai
httpx==0.25.0
gunicorn==21.2.0
psycopg2-binary==2.9.9